three pay channel

+++ wanglibao_pay/third_pay.py
@@ -8,6 +8,7 @@ sys.setdefaultencoding("utf-8")
 import logging
 from django.forms import model_to_dict
 from django.db import transaction
+from django.db.models import Q
 from django.utils.decorators import method_decorator
 from wanglibao_pay import util
 from wanglibao_pay.models import PayInfo, Bank, Card
@@ -21,7 +22,7 @@ from wanglibao_rest.utils import split_ua
 
 from wanglibao_pay.kuai_pay import KuaiPay
 from wanglibao_pay.huifu_pay import HuifuShortPay
-from wanglibao_pay.yee_pay import YeePay
+from wanglibao_pay.yee_pay import YeePay, YeeShortPay
 
 logger = logging.getLogger(__name__)
 
@@ -79,26 +80,21 @@ def list_bank_card(request):
     return {"ret_code":0, "message":"ok", "cards":rs}
 
 def del_bank_card(request):
+    """ 删除银行卡，需要解绑所有已绑定渠道"""
     card_id = request.DATA.get("card_id", "")
     if not card_id or not card_id.isdigit():
-        return {"ret_code":20041, "message":"请输入正确的ID"}
+        return {"ret_code": 20041, "message": "请输入正确的ID"}
 
-    card =  Card.objects.filter(id=card_id, user=request.user).first()
+    card = Card.objects.filter(id=card_id, user=request.user).first()
     if not card:
-        return {"ret_code":20042, "message":"该银行卡不存在"}
-    #删除快捷支付信息
-    storable_no = card.no[:6] + card.no[-4:]
-    pay = KuaiPay()
-    dic = {"user_id":request.user.id, "bank_id":card.bank.kuai_code,
-            "storable_no":storable_no}
+        return {"ret_code": 20042, "message": "该银行卡不存在"}
 
-    data = pay._sp_delbind_xml(dic)
-    res = pay._request(data, pay.DEL_URL)
-    logger.error("#api delete card")
-    logger.error(res.content)
+    # 删除快捷支付信息
+    res = _unbind_common(request, card, card.bank)
+    if res['ret_code'] != 0: return res
 
     card.delete()
-    return {"ret_code":0, "message":"删除成功"}
+    return {"ret_code": 0, "message": "删除成功"}
 
 def list_bank(request):
     #banks = Bank.get_deposit_banks()
@@ -196,14 +192,22 @@ def withdraw(request):
 
 def card_bind_list(request):
     # 查询已经绑定支付渠道的银行卡列表
-    from django.db.models import Q
-
     try:
         cards = Card.objects.filter(Q(user=request.user), Q(is_bind_huifu=True) | Q(is_bind_kuai=True) | Q(is_bind_yee=True)).select_related('bank').order_by('-last_update')
         if cards.exists():
+            # 排序
             bank_list = [card.bank.gate_id for card in cards]
             cards_tmp = sorted(cards, key=lambda x: bank_list.index(x['gate_id']))
             cards = cards_tmp
+
+            cards = [
+                {
+                    'bank_id': card.bank.id,
+                    'bank_name': card.bank.name,
+                    'gate_id': card.bank.gate_id,
+                    'storable_no': card[:6] + card[-4:]
+                } for card in cards]
+
             return {"ret_code": 0, "message": "ok", "cards": cards}
         else:
             return {"ret_code": 20031, "message": "请添加银行卡"}
@@ -221,7 +225,7 @@ def _unbind_kuaipay(request, card, bank=None):
 
 
 def _unbind_yeepay(request, card, bank=None):
-    return {"ret_code": 20104, "message": "未开通易宝支付渠道"}
+    return YeeShortPay().delete_bind(request.user, card, bank)
 
 
 def _unbind_common(request, card, bank):
@@ -264,26 +268,11 @@ def card_unbind(request):
     return _unbind_common(request, bank, card)
 
 
-def del_bank_card_new(request):
-    """ 删除银行卡，需要解绑所有已绑定渠道"""
-    card_id = request.DATA.get("card_id", "")
-    if not card_id or not card_id.isdigit():
-        return {"ret_code": 20041, "message": "请输入正确的ID"}
-
-    card = Card.objects.filter(id=card_id, user=request.user).first()
-    if not card:
-        return {"ret_code": 20042, "message": "该银行卡不存在"}
-
-    # 删除快捷支付信息
-    res = _unbind_common(request, card, card.bank)
-    if res['ret_code'] != 0: return res
-
-    card.delete()
-    return {"ret_code": 0, "message": "删除成功"}
-
-
-def bind_pay(request):
-    """ 根据银行设置的支付渠道进行支付渠道的支付 """
+def bind_pay_deposit(request):
+    """ 根据银行设置的支付渠道进行支付渠道的支付
+        1、获取验证码
+        2、快捷支付功能
+    """
     user = request.user
     card_no = request.DATA.get("card_no", "").strip()
 
@@ -302,9 +291,39 @@ def bind_pay(request):
         return HuifuShortPay().pre_pay(request)
 
     elif card.bank.channel == 'yeepay':
-        return {"ret_code": 20003, "message": "未开通易宝支付渠道"}
+        return YeeShortPay().pre_pay(request)
 
     elif card.bank.channel == 'kuaipay':
         return KuaiPay().pre_pay(request)
+    else:
+        return {"ret_code": 20004, "message": "请选择支付渠道"}
+
+
+def bind_pay_dynnum(request):
+    """ 根据银行设置的支付渠道进行支付渠道的支付
+        1、确认支付功能
+    """
+    user = request.user
+    card_no = request.DATA.get("card_no", "").strip()
+
+    if not card_no:
+        return {"ret_code": 20001, 'message': '信息输入不完整'}
+
+    if len(card_no) == 10:
+        card = Card.objects.filter(user=user, no__startswith=card_no[:6], no__endswith=card_no[-4:]).first()
+    else:
+        card = Card.objects.filter(no=card_no, user=user).first()
+
+    if not card:
+        return {"ret_code": 20002, "message": "银行卡未绑定"}
+
+    if card.bank.channel == 'huifu':
+        return {'ret_code': 20003, 'message': '汇付天下请选择快捷支付渠道'}
+
+    elif card.bank.channel == 'yeepay':
+        return YeeShortPay().dynnum_bind_pay(request)
+
+    elif card.bank.channel == 'kuaipay':
+        return KuaiPay().dynnum_bind_pay(request)
     else:
         return {"ret_code": 20004, "message": "请选择支付渠道"}
